+++ src/ts/ui/setTheme.ts
@@ -1,12 +1,7 @@
-import {setCodeTheme} from "../toolbar/CodeTheme";
-import {setContentTheme} from "../toolbar/ContentTheme";
-
-export const setTheme = (vditor: IVditor, codeTheme?: string) => {
+export const setTheme = (vditor: IVditor) => {
     if (vditor.options.theme === "dark") {
-        setContentTheme(vditor, "dark");
+        vditor.element.classList.add("vditor--dark");
     } else {
         vditor.element.classList.remove("vditor--dark");
-        setContentTheme(vditor, "light");
     }
-    setCodeTheme(vditor, codeTheme);
 };